ISSUE-794: Merge JsonSchemaUtil and JsonUtil
